put event related into events directory, separate logic
