--- updates/posted_events.json
@@ -1,1 +1,0 @@
-[1184165902019727371, 1186503466688979046, 1186504644629889085]
